pyproject.toml
[project]
name = "ed-auto-mission"
version = "2.0.0"
description = "Automatically accept missions in Elite Dangerous"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "congenial-acorn"},
]
keywords = ["elite-dangerous", "automation", "gaming"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Games/Entertainment",
]

dependencies = [
    "pyautogui>=0.9.50",
    "pydirectinput>=1.0.4",
    "numpy>=1.20.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "psutil>=5.9.0",
    "pywin32>=300; sys_platform == 'win32'",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
ed-auto-mission = "ed_auto_mission.main:main"

[project.urls]
Homepage = "https://github.com/Tropingenie/EDAutoMission"
Repository = "https://github.com/Tropingenie/EDAutoMission"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ed_auto_mission"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["ed_auto_mission"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false  # Start permissive, tighten later
ignore_missing_imports = true  # For pyautogui, pydirectinput, etc.

[[tool.mypy.overrides]]
module = [
    "pyautogui.*",
    "pydirectinput.*",
    "pytesseract.*",
    "win32gui.*",
    "cv2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
